app/requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
Django==3.2
django-cors-headers==4.0.0
django-split-settings==1.2.0
djangorestframework==3.14.0
isort==5.12.0
loguru==0.7.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
uWSGI==2.0.21
